last updated schema

+++ db/schema/02_cars.sql
@@ -3,7 +3,7 @@ DROP TABLE IF EXISTS cars CASCADE;
 CREATE TABLE cars (
     id SERIAL PRIMARY KEY,
     model VARCHAR(255) NOT NULL,
-    price INTEGER NOT NULL,
+    price FLOAT NOT NULL,
     description TEXT NOT NULL,
     image VARCHAR(255),
     date_created TIMESTAMP NOT NULL DEFAULT NOW(),
